+++ packages/supabase_codegen/lib/src/generator/generate_types/generate_types.dart
@@ -1,4 +1,6 @@
 export 'cmd_option.dart';
+export 'dart_type.dart';
 export 'default_values.dart';
 export 'get_config.dart';
 export 'run_generate_types.dart';
+export 'schema_overrides/schema_overrides.dart';