+++ packages/supabase_codegen/test/src/generators/generators/expected/row_fields.dart
@@ -67,8 +67,8 @@ const expectedRowFields = '''
   static const String isJsonField = 'is_json';
 
   /// Is Json
-  Map<String, dynamic>? get isJson => getField<Map<String, dynamic>>(isJsonField);
-  set isJson(Map<String, dynamic>? value) => setField<Map<String, dynamic>>(isJsonField, value);
+  dynamic get isJson => getField<dynamic>(isJsonField);
+  set isJson(dynamic value) => setField<dynamic>(isJsonField, value);
 
   /// Status field name
   static const String statusField = 'status';
@@ -76,5 +76,12 @@ const expectedRowFields = '''
   /// Status
   Status? get status => getField<Status>(statusField, enumValues: Status.values);
   set status(Status? value) => setField<Status>(statusField, value);
+
+  /// Is Dynamic field name
+  static const String isDynamicField = 'is_dynamic';
+
+  /// Is Dynamic
+  dynamic get isDynamic => getField<dynamic>(isDynamicField);
+  set isDynamic(dynamic value) => setField<dynamic>(isDynamicField, value);
 
 ''';